Fixed bug in AssignmentGraph.solver.   It did not create its result matrix correctly because it performed an incorrect test to check if the tail node is a reviewer and the head node is a paper.
To reveal this error I wrote 1 new (smaller) end_to_end test that exposed the flaw.  This meant a little refactoring of how I control silencing of print outs so I can control it from each individual test.   I also eliminated the use of random scoring in the metadata of
test cases because I suspected it might have been playing a role in the erroneous solutions of the matcher.  Test 8 does succeed as non-random scores where Test 8 fails with random scores (will address this later).  The test Params can now control whether scoring is random or fixed.

Finally, I built 3 unit-tests for the matchers AssignmentGraph solve algorithm using the SimpleGraphBuilder.  This isolated the bug in the solve method returning the incorrect matrix.   Made 3 such unit tests to try make sure we are correctly wrapping the google OR algorithm and working with its results.